final details
