+++ lib/calculator.rb
@@ -1,0 +1,11 @@
+module Calculator
+  def average(array)
+    array.inject{ |sum, element| sum + element }.to_f / array.count
+  end
+
+  def find_standard_deviation(array)
+    mean        = average(array)
+    sum_squares = array.inject(0) { |sum, element| sum + (mean - element)**2 }
+    Math.sqrt(sum_squares / (array.length - 1))
+  end
+end
