+++ lib/sales_engine.rb
@@ -15,12 +15,12 @@ class SalesEngine
               :customers
 
   def initialize(hash)
-    @items         =  ItemRepository.new(hash[:items], self)
-    @merchants     =  MerchantRepository.new(hash[:merchants], self)
-    @invoices      =  InvoiceRepository.new(hash[:invoices], self)
-    @invoice_items =  InvoiceItemRepository.new(hash[:invoice_items], self)
-    @transactions  =  TransactionRepository.new(hash[:transactions], self)
-    @customers     =  CustomerRepository.new(hash[:customers], self)
+    @items         ||=  ItemRepository.new(hash[:items], self)
+    @merchants     ||=  MerchantRepository.new(hash[:merchants], self)
+    @invoices      ||=  InvoiceRepository.new(hash[:invoices], self)
+    @invoice_items ||=  InvoiceItemRepository.new(hash[:invoice_items], self)
+    @transactions  ||=  TransactionRepository.new(hash[:transactions], self)
+    @customers     ||=  CustomerRepository.new(hash[:customers], self)
   end
 
   def self.from_csv(hash)
